kubelet: Force deleted pods can fail to move out of terminating
If a CRI error occurs during the terminating phase after a pod is
force deleted (API or static) then the housekeeping loop will not
deliver updates to the pod worker which prevents the pod's state
machine from progressing. The pod will remain in the terminating
phase but no further attempts to terminate or cleanup will occur
until the kubelet is restarted.

Alter the housekeeping loop to always deliver updates to the pod
worker when a pod is no longer in the config list, but only request
aggressive termination when the pod is unknown to the pod worker,
is not yet started, or has already terminated.

Refers to https://github.com/kubernetes/kubernetes/pull/113145

Manual cherry-pick of 45f626378b8c59fe4a83394e0c0e01b67f12cd33

Signed-off-by: Sascha Grunert <sgrunert@redhat.com>